backend/requirements.txt
# Core Flask dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# OpenAI and AI dependencies
openai>=1.30.0
whisper==1.1.10

# Computer Vision and Emotion Analysis
opencv-python==4.8.0.74
deepface==0.0.79
mediapipe==0.10.7
dlib==19.24.2
face-recognition==1.3.0

# Machine Learning and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
gtts==2.4.0

# Text Processing and NLP
textstat==0.7.3
phonetics==1.0.5
nltk==3.8.1
spacy==3.7.2

# Image Processing
pillow==10.0.0
imageio==2.31.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# Additional AI/ML libraries
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
tensorflow==2.13.0

# Development and Testing
pytest==7.4.2
black==23.9.1
flake8==6.1.0
